% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bike}
\alias{bike}
\title{Capital Bikeshare hourly demand data}
\format{
A data.frame with roughly 17,380 rows and the following columns:
\describe{
  \item{season, mnth, hr, weekday, weathersit}{Temporal and weather covariates.}
  \item{temp, hum, windspeed}{Numeric weather covariates.}
  \item{cnt}{Response: count of total rental bikes for that hour.}
}
}
\source{
UCI Machine Learning Repository:
  \url{https://archive.ics.uci.edu/dataset/275/bike+sharing+dataset}
}
\usage{
bike
}
\description{
Hourly count of rental bikes between 2011 and 2012 in the Capital Bikeshare
system with the corresponding weather and seasonal information. Used as the
regression demonstration in Srinivasan, Currim, and Ram (2025).
}
\keyword{datasets}
